Price Fix. Float Now

+++ database/migrations/2022_05_15_062511_create_cards_table.php
@@ -20,7 +20,7 @@ class CreateCardsTable extends Migration
             $table->string('city', 50);
             $table->string('street');
             $table->string('zip', 10);
-            $table->integer('price');
+            $table->float('price', 10, 2);
             $table->text('description');
             $table->biginteger('user_id')->unsigned();
             $table->biginteger('category_id')->unsigned();
